+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ## [Unreleased]
 ### Added
 - Pluggable flag-level help text rendering via `cli.DefaultFlagStringFunc`
+- `context.GlobalBoolT` was added as an analogue to `context.GlobalBool`
+- Support for hiding commands by setting `Hidden: true` -- this will hide the
+  commands in help output
 
 ### Changed
 - `Float64Flag`, `IntFlag`, and `DurationFlag` default values are no longer